Refacto getServerSession
Maintenant utilisé seulement dans withAuthentication/withServerSession et le helper pour les routes handleRouteErrors
