+++ src/pages/admin.tsx
@@ -2,7 +2,7 @@ import BulkEligibility from '@/components/Admin/BulkEligibility';
 import UserImpersonation from '@/components/Admin/UserImpersonation';
 import Users from '@/components/Admin/Users';
 import SimplePage from '@/components/shared/page/SimplePage';
-import { withAuthentication } from '@/server/helpers/ssr/withAuthentication';
+import { withAuthentication } from '@/server/authentication';
 
 export default function AdminPage(): JSX.Element {
   return (
