+++ src/pages/connexion.tsx
@@ -1,9 +1,6 @@
-import { type GetServerSideProps } from 'next';
-import { getServerSession } from 'next-auth';
-
 import { LoginForm, type LoginFormProps } from '@/components/connexion/LoginForm';
 import SimplePage from '@/components/shared/page/SimplePage';
-import { nextAuthOptions } from '@/pages/api/auth/[...nextauth]';
+import { withServerSession } from '@/server/authentication';
 
 export default function ConnectionPage(props: LoginFormProps): JSX.Element {
   return (
@@ -13,10 +10,8 @@ export default function ConnectionPage(props: LoginFormProps): JSX.Element {
   );
 }
 
-export const getServerSideProps: GetServerSideProps<LoginFormProps> = async (context) => {
-  const userSession = await getServerSession(context.req, context.res, nextAuthOptions);
-
-  if (userSession) {
+export const getServerSideProps = withServerSession(({ context, session }) => {
+  if (session) {
     return {
       redirect: {
         destination: '/gestionnaire',
@@ -30,4 +25,4 @@ export const getServerSideProps: GetServerSideProps<LoginFormProps> = async (con
       callbackUrl: (context.query.callbackUrl as string) || '/gestionnaire',
     },
   };
-};
+});
